+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Changelog
 **Unreleased**
 --------------
 
+- **Enhancement**: Always check for available assisted factories when reporting `InvalidBinding` errors about misused assisted injects.
+- **Enhancement**: Always specifically report mismatched assisted parameter mismatches.
+- **Enhancement**: Validate `Lazy` assisted factory injections in more places.
+- **Enhancement**: Allow private `@Binds` properties.
+- **Fix**: Register `MetroDiagnostics` in FIR.
+- **Fix**: When transforming FIR override statuses, check all supertypes and not just immediate supertype.
+
 0.6.6
 -----